type:   (NEW)  Original File Still Needs Editing
body
This program is the first attempt a Readme file that describes the 
program in detail.  This file should be the starting point for the user 
in analyzing the bikeshare.py dataset.   This file lays out the format 
that the final Readme file should have.  

the footer:  
See #26 and #9 (to edit more)